move some things into Tools
